requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.8.0
h11==0.14.0
idna==3.4
numpy==1.23.5
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
platformdirs==2.5.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.16.6
wsproto==1.2.0
